slight refactor to cleanup OptimizelyManager
