+++ OptimizelySDK/Optimizely/OptimizelyManager.swift
@@ -636,73 +636,6 @@ open class OptimizelyManager: NSObject {
     
 }
 
-extension HandlerRegistryService {
-    func injectLogger(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTLogger? {
-        return injectComponent(service: OPTLogger.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTLogger?
-    }
-
-    func injectNotificationCenter(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTNotificationCenter? {
-        return injectComponent(service: OPTNotificationCenter.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTNotificationCenter?
-    }
-    func injectDecisionService(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTDecisionService? {
-        return injectComponent(service: OPTDecisionService.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTDecisionService?
-    }
-    func injectBucketer(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTBucketer? {
-        return injectComponent(service: OPTBucketer.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTBucketer?
-    }
-    
-    func injectEventDispatcher(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTEventDispatcher? {
-        return injectComponent(service: OPTEventDispatcher.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTEventDispatcher?
-    }
-    
-    func injectDatafileHandler(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTDatafileHandler? {
-        return injectComponent(service: OPTDatafileHandler.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTDatafileHandler?
-    }
-    
-    func injectUserProfileService(sdkKey:String? = nil, isReintialize:Bool=false) -> OPTUserProfileService? {
-        return injectComponent(service: OPTUserProfileService.self, sdkKey: sdkKey, isReintialize: isReintialize) as! OPTUserProfileService?
-    }
-
-}
-extension OptimizelyManager {
-    func registerServices(sdkKey:String,
-                          logger:OPTLogger,
-                          eventDispatcher:OPTEventDispatcher,
-                          userProfileService:OPTUserProfileService,
-                          datafileHandler:OPTDatafileHandler,
-                          bucketer:OPTBucketer,
-                          decisionService:OPTDecisionService,
-                          notificationCenter:OPTNotificationCenter) {
-        // bind it as a non-singleton.  so, we will create an instance anytime injected.
-        // we don't associate the logger with a sdkKey at this time because not all components are sdkKey specific.
-        let binder:Binder = Binder<OPTLogger>(service: OPTLogger.self).to(factory: type(of:logger).init)
-        //Register my logger service.
-        HandlerRegistryService.shared.registerBinding(binder: binder)
-
-        // this is bound a reusable singleton. so, if we re-initalize, we will keep this.
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTNotificationCenter>(service: OPTNotificationCenter.self).singetlon().reInitializeStrategy(strategy: .reUse).using(instance:notificationCenter).sdkKey(key: sdkKey))
-
-        // this is a singleton but it has a reIntializeStrategy of reCreate.  So, we create a new
-        // instance on re-initialize.
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTBucketer>(service: OPTBucketer.self).singetlon().using(instance:bucketer).sdkKey(key: sdkKey))
-
-        // the decision service is also a singleton that will reCreate on re-initalize
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTDecisionService>(service: OPTDecisionService.self).singetlon().using(instance:decisionService).sdkKey(key: sdkKey))
-        
-        // An event dispatcher.  We rely on the factory to create and mantain. Again, recreate on re-initalize.
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTEventDispatcher>(service: OPTEventDispatcher.self).singetlon().reInitializeStrategy(strategy: .reUse).using(instance: eventDispatcher).sdkKey(key: sdkKey))
-        
-        // This is a singleton and might be a good candidate for reuse.  The handler supports mulitple
-        // sdk keys without having to be created for every key.
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTDatafileHandler>(service: OPTDatafileHandler.self).singetlon().reInitializeStrategy(strategy: .reUse).to(factory: type(of:datafileHandler).init).using(instance: datafileHandler).sdkKey(key: sdkKey))
-
-        // the user profile service is also a singleton using eh passed in version.
-        HandlerRegistryService.shared.registerBinding(binder:Binder<OPTUserProfileService>(service: OPTUserProfileService.self).singetlon().reInitializeStrategy(strategy:.reUse).using(instance:userProfileService).to(factory: type(of:userProfileService).init).sdkKey(key: sdkKey))
-
-    }
-}
-
-
 // MARK: - Objective-C Wrappers (WIP)
 extension OptimizelyManager {
